pyproject.toml
[project]
name = "recipe-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "litellm>=1.71.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.2",
]
